Combat UI State machine
Disabling "view" when a player has fully done an action, then resuming
Normal_Time_Flow, meaning actionbars values increase, waiting for input.

Have a problem with events not always registering, need to see what is
going wrong. Particularly when a player has finished a Combat UI
interaction after selecting a target. Sometimes the UI is disabled as
intended, sometimes it is not. Seems like it is not following the code /
a Dispatch not being fired even though it is in the code / a listener
not recieving correct object data when the Dispatch is fired.

Debug time..